Router and storemodulefactory
For next time:
- It does not fetch, why?
- If i only update with the generic setter, it does not refresh nor push. why?

- It uses the new router in services
- every domain has a file with the routes
- The books store now uses the storemodule factory
- Same for the authors store, but does not work properly

+++ app/Http/Controllers/BookController.php
@@ -73,9 +73,13 @@ class BookController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Book $book)
+    public function destroy($id)
     {
-        $book->delete();
+        $book = Book::find($id);
+        if ($book) {
+            $book->delete();
+        }
+
         $books = Book::with('reviews')->orderBy('created_at', 'desc')->get();
         return BookResource::collection($books);
     }
